Add PipelineContextDefinition
This adds the ability for pipeline to define all the contexts that it
supports. One can pass arguments to this context via the config. This
will pipelines to define contexts such as "debug" and "prod" as well as
others that are made available to all nodes in the pipeline. This is
excellent for concepts like handles to s3 and database connections.