Introduce resource pattern to the dagster sql context
This introduces the mixin-esque strategy to manage contexts. Instead
of subclassing per resource (e.g. A sqlalchemy-specific context), this will allow
a context to be utilize a number of different resources. Libraries will
assume that a resource of a certain name and type exist. Right now
the sqlalchemy kernel just does runtime type checks to enforce this.