Define selection execution
This begins to layout an approach for partial pipeline execution
in a world without sources.

The primary mechanism is through "create_pipeline_slice" which
allows the use to specify from_solids and through_solids to define
the subgraph, and then also pass in "injected_solids" where you specific
a dictionary just like dependencies, but instead of
DependencyDefinitions you just supply solids directly. If we can
build solids directly from a string (through some sort of registry), all
elements of this function call can be expressed from a config file.