Add RepositoryDefinition
Before we had no way of grouping pipelines together aside from the
pipelines.yml file. Instead, this moves that concept to a Python API
called the repository. A repository is a collection of pipelines.

Instead of specifying every pipeline in a project individually in a yaml
file, we now just have a single repository.

The repository object creates pipelines on demand (so that they can be
loaded by name without loading all the pipelines).

Repositories *also* have the concept of library solids (which can be
thought of higher order solids) which are ways to create solids in a
config-driven way. Authors will be able to design library solids that
are resuable. This will enable a layer of solid creation via config, and
much easier testing because authors will be able to inject pre-made
solids into existing pipelines to stub out values.