Clean up logging and fix double writing.
There are some concepts, some ideas, that are so repugnant, so vile, so debased,
that the world, by acclamation, that unite the world with a single voice
shouting "Never Again." Some things should be tossed into history's unmarked
grave of discarded lies. To this list I nominate global, mutable variables.
Against my better judgement I used a global set to deal with python's logging
global-ness. And naturally it is the source of the first real, honest-to-God
bug in production.

Now that I'm more confident about the way the context I am just going to
totally avoid python's horrific global logger registry and just create
loggers directly. This ends up being a lot easier to think about,
manage, and test.