Add a manager notion to dagstermill
This a bit unfortunate but necessary. Interactions with dagstermill will
got through a manager object. This incurs some conceptual weight on the
API consumer, but allows for greater infrastructure flexibility and the
removal of some global state.

The real motivation behind this (hence the test to include a solid
from within a notebook and moving the tests to within a module) is
to allow the notebook to include the SolidDefinition the defines it.
This will enable the control of the serialization and deserialization
process using the dagster type system as the guide.