No longer have separate flags for pipeline and repo functions (#208)
Instead of having -p and -r flags for pipelines and repos functions
for cli commands that target a pipeline, we just have a single flag -n.
Whether or not it attempts to target a repo or pipeline depends on the
presence of a name argument. This also allows for a unified interface
between dagster and dagit on this front.