Update to version 0.2.4 (#223)
This version bump contains a few changes (including one breaking
change).

- New, radically improved version of dagit. Vertical layout, and a
beautiful new design. H/T to @bengotow for this spectacular work.
- All types now *require* names. This is breaking change for
ConfigDictionary, which did not require a name. You will
have to change your calls to ConfigDictionary or
ConfigDefinition.config_dict to include a name that is unique to the
Pipeline.
- Solids default to take *no* config definition, rather than a config
definition typed as any.