Better rules around what you need to specify for context configs (#250)
Context configuration is complicated. The current system config schema is working around the lack of something like tagged union. Only one of the contexts can be specified, and 'default' is treated specially. (The default thing might definitely be a mistake, but we can address that later).

I expect this to change as we explore the config editor.