Improve hot-reloading, display Python errors in Dagit UI (#242)
* Capture, expose most recent python error to GraphQL

* Simplify implementation of dynamic loading, handle case where initial load fails

The problem with the current implementation of DynamicObject is that the tuple contains both the `module` reference and also the evaluated result. If you can’t evaluate the function, you can’t obtain a DynamicObject, which means you have no module handle to use to call `importlib.reload`.

To fix this I’ve split the loading into a two phase process—we create a `DynamicObject` with just the module and function name, and then evaluate the module separately. This allows 1) the evaluation to be done in the try catch and 2) means the initial load can fail and still have a DynamicObject with a module reference. I also made `eval` idempotent so it can also replace `reload_dynamic_object`.

* Display errors in the React UI

* Fix the tests

* Fix tests, allow RepositoryContainer(repository=obj) for inline repos

* Put the new error field in the mock graphql response to fix JS test

* DynamicLoader eval => load

* Very slightly move code coercing pipeline to repo

* Expose pipelinesOrErrors, pipelineOrError at top level of graphql schema

* Client-side updates to use pipelinesOrErrors

* Use the reloader module to build a dependency graph, reload everything in order

* Fix the web app mock query data

* Do a bit of linting

* Handle non-pipeline, non-repository case with fall through exception

* Included reloader in the wrong python setup.py

* On Python 2, disable module reloading by stubbing `reloader`

* Only install reloader on python3, setup script isn’t even compatible with 2.x

* Reload dagster itself (though it is not in the watched path)

* check.inst_param => check.inst

* Just throw rather than nicely returning if no repository is present

* Revert "Very slightly move code coercing pipeline to repo"

This reverts commit 24548f69c406a7d2aa0e10ba1807edf637616b44.

# Conflicts:
#	python_modules/dagster/dagster/cli/dynamic_loader.py

* Explicitly check for two supported types

* Fix tests

* Add comment