Better management of ConfigDictionaries (#255)
 Create methods for easily constructed nested configs on solids and
contexts

Right now, constructing nested dictionaries for configurations on solids
and contexts is a big pain. The user must make up a throwaway name for
type and then do multiple layers of nesting, coming up with more
arbitrary names a long the way. The syntax is also verbose.

This PR address that issue.

What was before:

   ConfigDefinition(
        types.ConfigDictionary(
            'PipelineName.Solid.SolidName.ConfigDict',
            {
                'foo' : types.Field(types.String),
                'nested_dict' : types.Field(
                    types.ConfigDictionary(
                        'PipelineName.Solid.SolidName.NestedDict.ConfigDict',
                        {
                            'bar' : types.Field(types.String),
                        },
                    ),
                ),
            },
        ),
    )

becomes

    ConfigDefinition.solid_config_def_dict(
        'pipeline_name',
        'solid_name',
        {
            'foo': types.Field(types.String),
            'nested_dict': {
                'bar': types.Field(types.String),
            },
        },
    )

This should make it far less painful to construct configuration types
that are custom to a context or solid, reducing lots of boilerplate.