Expose GraphQL Interface for Config Errors (#285)
This adds a GraphQL interface for detecting errors in the config, and reports all the errors that are present in the config document. This will make it possible to add a richer editing experience to the config editor.

This diff also fixes the major oversight that I never actually applied the new "selector" code path to the context configuration that it was designed for!