Move Execution Plan query to root (#309)
Mikhail and I managed to screwed the merging of some stacked diffs a couple nights ago. This is recovering that change so I can build on the code while developing the pipeline execution codepaths.

The apollo type generator is being a bit weird here in the presence of union queries with interface fallbacks, so I changed it to just check for all the concrete types in the union. This is worthy of discussion/review.

However I'm going to move forward with merging this to master. We can change the client code later.