Config system improvements (including better path information) (#330)
This turned into a much bigger PR than I would have liked. However it takes care of some issues which have been plaguing the system for a bit.

execute_pipeline only takes a dict, not a config.Environment. This is the reason for most of the churn here, as a ton of unit tests needed updated. However this allows for more clear error messages, simplified code, and unit tests more in line with how we want all client code to be written.
Internal refactoring of the type system. A lot of the internal mechanisms of the config type system were rewritten to be more generic.
Much better error messages. We now print out good path information for errors . E.g. “Missing required field "config" at path root:solids:good_error_handling”
This isn't quite where I want it to be but this is a good incremental stopping point. It may be worth shipping down these error messages down to the dagit as well.