Output Materializations (#479)
This creates a new system for specifying materializations of outputs via config. In any config file, a user can specify multiple materializations (both for multiple outputs and multiple materializations for a single output). This is why the config schema is a list.

A user creates a materializable type by implementing Materializeable. You must specify your config schema and implement a function that does the actual materialization.

Definitely a few unsolved questions in here such as how these interact with resources and how one can mixin new materialization schemas with existing types. However I think this is a good start.

This PR should also serve as a nice example of how to add a feature to the execution plan subsystem.