New Dagstermill Marshalling (#486)
This considerably restructures dagstermill with a new API that requires less boilerplate (in exchange for introducting some global state) and allows for marshalling values between processes. We use this in this diff to marshall data frames to and from a notebook solid. This allows an author to define notebooks in terms of their abstract inputs and outputs.