Avoid infinite loops in type iterate (#541)
Well we finally are paying a real price for the confusion between the two type systems. This definitely demonstrates that we need a separate notion of scalars for both the config and runtime systems. The builtin-scalars contain a materialization schema which contain types that contain those scalars. We have to cut short the recursion at some point, and to do that we pass down a set of seen config schemas so we don't iterate over them twice.