Integrate process lifecycle events into the execution plan view (#550)
Addresses #530 

I originally implemented this as three separate messages in the execution plan view, but @mgasner suggested a single line which we append detail to, and I think that works really well. Now it looks like this (except the PIDs would match). The ellipsis in the sentence animate to make it super clear that something is happening.