Type System Rearchitecture: Separate Config and Runtime Types (#547)
Type system re-architecture.

This is fairly massive change to type system. Before we had a single type system that contained a complicated set of mixins to indicate whether or not you could specify whether or not that type's inputs and outputs could be specified in config. This was quickly becoming untenable and quite confusing. After a discussion with @bengotow last week I decided to take a stab at clean break between the config and runtime type systems.

The awkwardness had manifested itself a few ways:

1) The notion of a "Dict" was confusing because it really could describe a couple different things. One is our notion of strongly-typed dictionary that provides the typeahead and checking for config files; the other is the whether or not the value flowing at runtime is a python dictionary. These are truly orthogonal concepts and should be thought of as such.

2) When adding materialization options to scalars this got really nasty. In effect because the runtime and config type system were the same. If you had a String output of a solid, that in turn could be materialized to a file and you had to specify a string path for the file, without hardcoding some weird solution you would get into an infinite loop as you descended down the type tree. This really exposed the fact that the solids output "String" and the "String" entry in the materialization config schema are two separate concepts.

3) When I attempt to make so that you could "source" a runtime Int *either* from a file or from the value embedded within the config file itself, the same type of infinite looping resulted.

There are now two parallel type systems: one for runtime values; one for the config values. Each has their own "Int", "List" "Nullable" and so forth. **Only** the config system contains the notion of Composites and Selectors. Runtime types can specify that they can be instantiated via config by specifying input and output schema in the **config** type system.

Things remaining:

1) Right now in order to implement a RuntimeType you have to implement a parameterless __init__ method and pass up attributes in the super constructor. It is probably more "Pythonic" to do some Django/Graphene-style wizardry with metaclasses make it so the user specifies these attribute at the class level.

2) GraphQL and dagit. I did not change the GraphQL schema to account for this. Instead I wrote some hacky merging code to be faithful to the old semantics. We should expose runtimeTypes and configTypes fields separately and change the UI to match.


Some Benefits in no particular order:

- When declaring DagsterTypes you no longer have to go through the song and dance of declaring a class and the instantiating it and setting it to some global and using that. You just pass the class directly. This should essentially completely eliminate the "type uniqueness" usability issue.
- Have different types makes user extensions more clear than the crazy mixin system imo.
- Config representation and in-memory representation are now totally decoupled, as it always should have been.

In order to simplify the user experience, we hide the fact that there are two type systems via some wizard. See comment in dagster/core/types/__init__.py for gory details.

We are also going to move to a convention where we do *not* use the `types.` scoping for types. So `types.Int` becomes `Int`. For nested types this ends up being a surprisingly big usability wein.