New Input/Output Schema API (#611)
This is a new, leaner API for specifying input and output schemas on user-defined types. An input and output schema at its core is a config type and then a function that transforms that a process config value to an in-memory object compatibility with the runtime type the schema is applied to.

This expresses schema just a decorated function then. We a variant for any type and a lightly specialized variant for selector types, which is common in these contexts.

So far example the input schema for a pandas data frame is fully specified as:

```
@input_selector_schema(
    Selector(
        {
            'csv': define_csv_dict_field(),
            'parquet': define_path_dict_field(),
            'table': define_path_dict_field(),
        }
    )
)
def dataframe_input_schema(file_type, file_options):
    if file_type == 'csv':
        path = file_options['path']
        del file_options['path']
        return pd.read_csv(path, **file_options)
    elif file_type == 'parquet':
        return pd.read_parquet(file_options['path'])
    elif file_type == 'table':
        return pd.read_table(file_options['path'])
    else:
        check.failed('Unsupported file_type {file_type}'.format(file_type=file_type))
```