Remove legacy graphQL interface, move dagit to use configTypes / runtimeTypes (#723)
This PR removes the `legacy_types.py` file and with it the 'Type' and 'TypeField' types. This required a bit of work on the python side and I annotated one spot I think I may have done something questionable.

On the React front this was pretty straightforward thanks to TypeScript. The type explorer now only operates on runtime types and the config scaffolding / config autocompletion now uses the type `keys`.

A few caveats:
- You can't click the types within a config schema text block, since there is no explorer for config types.
- The `Pipeline Types` section of the type explorer only shows types that are in use and does not separate "Built In" types and "Custom" types the way it's meant to. For now I've totally removed the subsections. I think it should always show all the built ins, as well.

Addresses #589 and #575