Correctly display lists and nullables in the runtime type domain (#749)
Previous type system refactor had introduced a regression when input type and output types of solid were listed as "Unnamed Type". This adds a server side property to render [] and ? for lists and nullables. This is a bit of a stop gap. I think we will have to switch this to be client-side to get all the rendering goodness we want, but this is an improvement for now.