ExecutionContext Rearchitecture (#801)
This is a big PR that totally eliminates the old weird structure of the execution context and instead does a first principles approach that matches the public API.

The biggest material change here is that there is not a separate notion of a context object per use case, and they inherit from eachother.

So there is an ExpectationExecutionContex and TransformExecutionContext. They inherit from StepExecutionContext, which in turn inherits from PipelineExecutionContext. While this is quite verbose we also in a strongly typed fashion know what is available at any point during execution. No more accessing from the tags. Also still have the TransformExecutionContextMetadata ABC so that dagstermill can stub out its own version.

In order to maintain backwards compat we have a separate class LegacyContextAdapter which can be easily deleted once we break that API guarantee.