Exit properly on double Ctrl-C when running in watch mode (#930)
This PR replaces the bash script I wrote to execute `dagit-cli` or `watchmedo dagit-cli` with a Python script that uses watchdog's API directly.

The problem we observed with the other approach is that the watchmedo CLI attempts to clean up the file watchers BEFORE sending a signal to the subprocess it launches. This process takes a while, and sending a second SIGINT in this time window kills the watchmedo process without a signal ever getting sent to the subprocess.

The relevant code in watchdog is here: https://github.com/gorakhargosh/watchdog/blob/master/src/watchdog/watchmedo.py#L124. This function runs `observer.stop()`, `observer.join()`, and then in the callsite the `handler` is stopped which kills the subprocess.

Thankfully watchdog's python API is pretty nice and watchmedo uses it directly - I basically pulled 20 lines of watchmedo out into our launch script and added the `handler.stop()` call that is missing.